Change room row id`s from 'long' to 'int'
* Change primary and foreign key`s from 'long' to 'int' because is
  not necessary to use a 'long' value. An 'int' will be enough.

+++ app/src/main/java/com/smart_learn/core/services/LessonService.java
@@ -25,7 +25,7 @@ public class LessonService extends BasicRoomService<Lesson> {
         super.basicRoomRepository = lessonRepository;
     }
 
-    public LiveData<Lesson> getSampleLiveLesson(long lessonId) {
+    public LiveData<Lesson> getSampleLiveLesson(int lessonId) {
         return lessonRepository.getSampleLiveLesson(lessonId);
     }
