+++ app/src/main/java/com/smart_learn/data/models/room/entities/helpers/LessonEntrance.java
@@ -11,14 +11,14 @@ public abstract class LessonEntrance {
 
     private final long createdAt;
     private long modifiedAt;
-    private long fkLessonId;
+    private int fkLessonId;
     private boolean isSelected; // helper for recycler view
 
     /** Leave this embedded in order to be decomposed in room_db */
     @Embedded
     private Translation translation;
 
-    public LessonEntrance(long createdAt, long modifiedAt, long fkLessonId, boolean isSelected, Translation translation) {
+    public LessonEntrance(long createdAt, long modifiedAt, int fkLessonId, boolean isSelected, Translation translation) {
         this.createdAt = createdAt;
         this.modifiedAt = modifiedAt;
         this.fkLessonId = fkLessonId;
